% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qry_user.R
\name{add_user}
\alias{add_user}
\title{Add a user to the database}
\usage{
add_user(user_name, last_login, access_rights, testing = FALSE)
}
\arguments{
\item{user_name}{The user's name}

\item{last_login}{The user's last login}

\item{access_rights}{The user's access rights}

\item{testing}{Whether to run in testing mode, will rollback
the db transaction if TRUE}
}
\value{
true if successful
}
\description{
Add a user to the database
}
