% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{get_db_connection}
\alias{get_db_connection}
\title{Create database connection}
\usage{
get_db_connection()
}
\description{
Currently sqlite, please note that sqlite has significant limitations
in a multi-tenant environment with several modules sharing a database
and is not recommended for production use unless it fits your use case.
}
