% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qry_user.R
\name{get_users}
\alias{get_users}
\title{Retrieve all users in a df}
\usage{
get_users()
}
\value{
A data frame with user data
}
\description{
This retrieves user data from the database
(not always advisable to read the entire table)
}
